requirements.txt
django
tzdata
celery
requests
python-telegram-bot
asgiref